requirements.txt
langgraph
torch
pillow
diffusers
networkx
matplotlib
langchain
langchain_community
langchain_groq
arxiv
wikipedia
gradio
langsmith
transformers[agents]
git+https://github.com/huggingface/transformers.git
git+https://github.com/suno-ai/bark.git
soundfile
python-dotenv
